requirements.txt
streamlit
openai>=1.0.0
pandas
matplotlib
python-docx
openpyxl
